// bootstrap/compiler/src/generated/graph-capabilities.contracts.ts
export type GraphDocumentLowerAdapter = {
	lower: (input: ModelTypes.GraphDocumentLowerInput) => Promise<ModelTypes.GraphDocumentLowerResult>;
	prepare: (input: ModelTypes.GraphDocumentLowerPrepareInput) => Promise<ModelTypes.GraphDocumentLowerHandle>;
};
// ...
export type GraphProvideAdapter = {
	provide: (input: ModelTypes.GraphProviderInput) => Promise<ModelTypes.GraphProviderResult>;
};
// ...
export type GraphValidateAdapter = {
	validate: (input: ModelTypes.GraphValidateInput) => Promise<ModelTypes.GraphValidationResult>;
};